Bu release'de scheduled job seklinde github'a gidip 1 dakikada bir yeni issue varmi diye bakacagiz. Bunun icin spring'in anotasyonlarini ve async methodlari kullanacagiz.
configuration package'i altina Async methodlar icin kullanilacak bir
executor configuration'i barindiracak AsyncConfig isimli dosyasi create
ettim ve spring'e ait @EnableSync isimli anotasyonu kullandim.
Icerisinde ki bana executor donen method aslinda benim @Async ile
isaretledigim methodlari execute etmeye yarayacak. Main threadden haric
bir thread olusturup async calisacak. Bu mekanizmayi kullanabilmek icin
bir configuration sinifi create edilebilir ya da bu ornekte yaptigim
gibi Application.java adli baslangic class'i ma @EnabledScheduling
anotasyonu ile ben bu projede scheduling ile islemler yapacagim diye
belirtebilirim.

GithubProperties icerisine @Component anotasyonunu ekledim

RestTemplate uygulamayi ayaga kaldirirken bean hatasi veriyordu
configuration klasoru altina RestTemplateConfiguration isimli class'i
create ettim ve icerisine RestTemplate donduren bir bean tanimladim.

RepositoryManager class'im icerisin de issue listesi olustururken
kullandigim new ile olusturdugum issue nesnesi hatali oldugu icin
builder pattern'i kullanarak issueModel'i kullanarak builder yardimi ile
title'i ve body'si olan object'i create edip listemi o sekilde
olusturdum.

IssueModel isimli model class'im da body isimli String degiskenim icin
@Column anotasyonu ile bunun columnDefinition olarak text kullanacagimi
soyledim. Cunku gelen response'un icerisinde ki body cok buyuk

RepositoryModel icerisine createdAt ve updatedAt isimli date tipinde 2
field ekledim bu field'ler repository create edildigi anda otomatik
DB'de olusacak tarihleri belirtiyor olacaklar. @CreationTimeStamp ve
@UpdateTimestamp anotasyonlari yardimi ile

IssueRepository isminde ListCrudRepository'i extends eden ve otomatik
save yada listeleme methodlari icin JPA kullanan class'i create ettim

ImportIssuesScheduler isimli class ile application.yml icerisinde
yaratmis oldugum application tagina ait import-frequency degerini
60000ms olarak set edip (1 dk) ardindan ilk basta calismamasi icin
initialDelay parametresi olarak da 60000 ms degerini kullanip
issue'larin import edilmesini sagladim. Bu methodda @Scheduled
anotasyonu kullanildi.

GithubClientService icerisinde ki listIssues methodu imzasina bir
LocalDate formatinda since degeri ekledim. Bu deger sorguyu attigimiz
gunun her zaman -1 gununcu degerini String tipinde bu methoda verecek bu
methodda issue'lari ona gore response'u dondurecek

IssueService isimli service'i create ettim. Bu servis list olarak gelen
issue'lari DB'ye yazacak

GithubIssueResponse olarak issues.json dosyasinin map edildigi class'da
int olan id degerleri long'a cevirildi

application.yml altina application adinda bir tag eklendi. Onun da
icerisine import-frequency degeri olarak 60000 degeri atandi.

GithubClientServiceTest icerisin de ise since parametresine ait yeni
testler yazildi

+++ src/main/java/com/example/youcontribute/service/models/GithubIssueResponse.java
@@ -17,7 +17,7 @@ public class GithubIssueResponse {
     public String commentsUrl;
     public String eventsUrl;
     public String htmlUrl;
-    public int id;
+    public long id;
     public String nodeId;
     public int number;
     public String title;
@@ -56,7 +56,7 @@ public class GithubIssueResponse {
     }
 
     public static class Label {
-        public int id;
+        public long id;
         public String nodeId;
         public String url;
         public String name;
@@ -68,7 +68,7 @@ public class GithubIssueResponse {
 
     public static class Assignee {
         public String login;
-        public int id;
+        public long id;
         public String nodeId;
         public String avatarUrl;
         public String gravatarId;
@@ -89,7 +89,7 @@ public class GithubIssueResponse {
 
     public static class User {
         public String login;
-        public int id;
+        public long id;
         public String nodeId;
         public String avatarUrl;
         public String gravatarId;
